+++ Server/src/GameWorld/Player/Player.h
@@ -32,7 +32,7 @@ class Player {
         void stopMovementDirection(MoveTo direction);
         void setShootingState();
         void stopShootingState();
-        void applyStep(std::map<int, std::shared_ptr<Collidable>>& collidables,
+        void applyStep(std::map<uint8_t, std::shared_ptr<Collidable>>& collidables,
                        std::map<uint8_t, std::shared_ptr<Infected>>& infecteds);
         std::pair<int16_t, int16_t>& getPosition();
         TypeOperator& getTypeOperator();
@@ -40,7 +40,7 @@ class Player {
         uint8_t& getHealth();
         virtual ~Player() = default;
     private:
-        void move(std::map<int, std::shared_ptr<Collidable>>& collidables);
+        void move(std::map<uint8_t, std::shared_ptr<Collidable>>& collidables);
         void shoot(std::map<uint8_t, std::shared_ptr<Infected>>& infecteds);
         virtual void specialAtack(Event event) = 0;
 };
